# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Spector Cognitive Memory & Search API
  description: "Ultra-fast cognitive memory, vector search, and agent orchestration\
    \ platform with biological memory tiers (Working, Episodic, Semantic, Procedural)."
  contact:
    name: Spectrayan Maintainers
    url: https://github.com/spectrayan/spector
    email: support@spectrayan.com
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
  description: Current Server Instance
security:
- ApiKeyAuth: []
- BearerAuth: []
...
paths:
...
  /tasks/{id}/trigger:
    post:
      tags:
      - task-management-controller
      operationId: triggerTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /api/v1/tasks/{id}/trigger:
    post:
      tags:
      - task-management-controller
      operationId: triggerTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /tasks/{id}/resume:
    post:
      tags:
      - task-management-controller
      operationId: resumeTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /api/v1/tasks/{id}/resume:
    post:
      tags:
      - task-management-controller
      operationId: resumeTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /tasks/{id}/reschedule-interval:
    post:
      tags:
      - task-management-controller
      operationId: rescheduleInterval
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleIntervalRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /api/v1/tasks/{id}/reschedule-interval:
    post:
      tags:
      - task-management-controller
      operationId: rescheduleInterval_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleIntervalRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /api/v1/tasks/{id}/reschedule-cron:
    post:
      tags:
      - task-management-controller
      operationId: rescheduleCron
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleCronRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /tasks/{id}/reschedule-cron:
    post:
      tags:
      - task-management-controller
      operationId: rescheduleCron_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleCronRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /tasks/{id}/pause:
    post:
      tags:
      - task-management-controller
      operationId: pauseTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
  /api/v1/tasks/{id}/pause:
    post:
      tags:
      - task-management-controller
      operationId: pauseTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskActionResponse"
...
  /api/v1/tasks/{id}/audit:
    get:
      tags:
      - task-management-controller
      operationId: getTaskAuditHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRunAuditRecord"
  /tasks/{id}/audit:
    get:
      tags:
      - task-management-controller
      operationId: getTaskAuditHistory_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRunAuditRecord"
  /tasks/{id}:
    get:
      tags:
      - task-management-controller
      operationId: getTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskStatus"
  /api/v1/tasks/{id}:
    get:
      tags:
      - task-management-controller
      operationId: getTask_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskStatus"
  /tasks/audit:
    get:
      tags:
      - task-management-controller
      operationId: getRecentAuditHistory
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRunAuditRecord"
  /api/v1/tasks/audit:
    get:
      tags:
      - task-management-controller
      operationId: getRecentAuditHistory_1
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskRunAuditRecord"
...
components:
  schemas:
...
    TaskActionResponse:
      type: object
      properties:
        status:
          type: string
        taskId:
          type: string
        message:
          type: string
    RescheduleIntervalRequest:
      type: object
      properties:
        intervalSeconds:
          type: integer
          format: int64
    RescheduleCronRequest:
      type: object
      properties:
        cronExpression:
          type: string
...
    TaskRunAuditRecord:
      type: object
      properties:
        runId:
          type: string
        taskId:
          type: string
        namespaceId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: string
        status:
          type: string
        result: {}
        errorMessage:
          type: string
        success:
          type: boolean
    TaskStatus:
      type: object
      properties:
        id:
          type: string
        namespaceId:
          type: string
        state:
          type: string
        previousFireTime:
          type: string
          format: date-time
        nextFireTime:
          type: string
          format: date-time
        description:
          type: string
        metadata:
          type: object
          additionalProperties: {}
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Spector API Key Authentication
      name: X-API-Key
      in: header
    BearerAuth:
      type: http
      description: OAuth2 / OIDC JWT Bearer Token
      scheme: bearer
      bearerFormat: JWT